+++ mitra_activitypub/src/actors/attachments.rs
@@ -72,7 +72,9 @@ pub fn parse_identity_proof_fep_c390(
     let signer = match signature_data.verification_method {
         VerificationMethod::HttpUrl(_) =>
             return Err(ValidationError("unsupported verification method")),
-        VerificationMethod::DidUrl(did) => did,
+        // Fragment is ignored because supported DIDs
+        // can't have more than one verification method
+        VerificationMethod::DidUrl(did_url) => did_url.did().clone(),
     };
     if signer != statement.subject {
         return Err(ValidationError("subject mismatch"));