+++ apx_sdk/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Return `Response` from `send_object` instead of `Option<Response>`.
 - Allow choosing backend for `idna`.
 - Improved documentation.
+- Keep full `DidUrl` when creating `VerificationMethod::DidUrl` enum variant.
 
 ### Removed